// src/app/common/models/40sys_member.models.ts
export class Sys_member {
    id?: Number;
    maphongban?: string;
    manhom?: string;
    kyvanban?: string;
    name?: string;
    siteid?: string;
    password?: string;
    username?: string;
    postdate?: string;
    approved?: any;
    level?: string;
    permission?: string;
    email?: string;
}